+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ---
 
+## [0.15.0](https://github.com/FortAwesome/angular-fontawesome/releases/tag/0.15.0)
+
+Make sure to check [upgrade instructions](https://github.com/FortAwesome/angular-fontawesome/blob/master/UPGRADING.md).
+
+### Removed
+
+- Deprecated inputs `spin` and `pulse`. Use `animation` input instead.
+
 ## [0.14.1](https://github.com/FortAwesome/angular-fontawesome/releases/tag/0.14.1) - 2023-12-23
 
 ### Fixed